pyproject.toml
[tool.poetry]
name = "Project-X-Backend"
version = "0.1.0"
description = "Project X Backend SVC"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.15"
langchain-openai = "^0.1.3"
python-dotenv = "^1.0.0"
openai = "^1.7.0"
peewee = "3.13.1"
redis = "^5.0.1"
pymysql = "0.9.3"
networkx = "^3.2.1"
simplejson = "^3.19.2"
aenum = "^3.1.15"
numpy = "^1.26.3"
falcon = "^3.1.3"
gevent = "23.9.1"
falcon-marshmallow = "0.3.0"
marshmallow = "3.13.0"
marshmallow-union = "^0.1.15.post1"
pydantic = "^2.6.1"
boto3 = "^1.34.38"
pillow = "^10.3.0"
pyjwt = "^2.8.0"
pika = "^1.3.2"
streamlit = "^1.34.0"
streamlit-extras = "^0.4.2"
ftfy = "^6.2.0"
retry = "^0.9.2"
joblib = "^1.4.2"
wheel = "^0.43.0"
google-cloud-storage = "^2.18.2"
pusher = "^3.3.2"
langchain-core = "^0.2.38"
google-cloud-translate = "^3.16.0"
langchain-google-vertexai = "^1.0.10"
google = "^3.0.0"
langchain-community = "^0.2.16"
pdfminer-six = "^20240706"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-dotenv = "^0.5.2"
debugpy = "^1.8.1"
ragas = "^0.1.16"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
